fix: improve the cascading effect of delete
Delete occurs at model method and queryset method

Also, deleting attribute or anchor have different side effects.

Deleting anchor also requires cascading the attributes to be deleted

Deleting anything with transaction and historized requires the historized to be updated correctly.

Deleting is harder than updating because the instance is already removed, so special care is required to prepare the required data prior to the actual delete.

Else it’s hard to propagate downstream changes due to the delete.

Signed-off-by: KimSia, Sim <245021+simkimsia@users.noreply.github.com>